+++ packages/trimerge-sync-indexed-db/src/trimerge-indexed-db.ts
@@ -38,45 +38,6 @@ function toSyncNumber(syncId: string | undefined): number {
   return syncId === undefined ? 0 : parseInt(syncId, 36);
 }
 
-/** This creates a native browser BroadcastChannel implementation of EventChannel */
-function getBroadcastChannelEventChannel<CommitMetadata, Delta, Presence>(
-  docId: string,
-): EventChannel<CommitMetadata, Delta, Presence> {
-  let channel: BroadcastChannel | undefined = new BroadcastChannel(docId);
-  const eventListenerCallbacks: ((e: MessageEvent) => void)[] = [];
-
-  return {
-    onEvent: (
-      cb: (ev: BroadcastEvent<CommitMetadata, Delta, Presence>) => void,
-    ) => {
-      if (!channel) {
-        throw new Error(
-          'attempting to register an event callback after channel has been shutdown',
-        );
-      }
-
-      const newCb = (e: MessageEvent) => cb(e.data);
-      eventListenerCallbacks.push(newCb);
-      return channel?.addEventListener('message', newCb);
-    },
-    sendEvent: (ev: BroadcastEvent<CommitMetadata, Delta, Presence>) => {
-      if (!channel) {
-        throw new Error(
-          `attempting to send an event after channel has been shutdown ${ev}`,
-        );
-      }
-      return channel?.postMessage(ev);
-    },
-    shutdown: () => {
-      for (const cb of eventListenerCallbacks) {
-        channel?.removeEventListener('message', cb);
-      }
-      channel?.close();
-      channel = undefined;
-    },
-  };
-}
-
 type LocalIdGeneratorFn = () => Promise<string> | string;
 export type IndexedDbBackendOptions<CommitMetadata, Delta, Presence> = {
   getRemote?: GetRemoteFn<CommitMetadata, Delta, Presence>;
