+++ packages/trimerge-sync/src/index.ts
@@ -4,4 +4,5 @@ export * from './types';
 export * from './AbstractLocalStore';
 export * from './validateCommits';
 export * from './lib/PromiseQueue';
+export * from './lib/EventChannel';
 export * from './merge-all-helper';
