+++ packages/trimerge-sync/src/lib/EventChannel.ts
@@ -1,0 +1,16 @@
+import { SyncEvent } from '../types';
+
+export type BroadcastEvent<CommitMetadata, Delta, Presence> = {
+  event: SyncEvent<CommitMetadata, Delta, Presence>;
+  remoteOrigin: boolean;
+};
+
+export type EventChannel<CommitMetadata, Delta, Presence> = {
+  onEvent(
+    cb: (ev: BroadcastEvent<CommitMetadata, Delta, Presence>) => void,
+  ): void;
+  sendEvent(
+    ev: BroadcastEvent<CommitMetadata, Delta, Presence>,
+  ): void | Promise<void>;
+  shutdown(): void | Promise<void>;
+};
